+++ core/lib/sn/CmnInterface.d.ts
@@ -116,6 +116,7 @@ export declare type HArg = {
     end_ms?: number;
     join?: boolean;
     do_rec?: boolean;
+    stepin?: boolean;
     ':id_dc'?: string;
     ':id_tag'?: string;
     ':path'?: string;
@@ -125,7 +126,6 @@ export declare type HArg = {
     ':idx_tkn'?: number;
     ':token'?: string;
     design_unit?: string;
-    design?: boolean;
 };
 export interface ITag {
     (hArg: HArg): boolean;